minitalk

+++ 02_minitalk/minitalk/server/minitalk_server.c
@@ -6,7 +6,7 @@
 /*   By: jiychoi <jiychoi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/17 15:27:39 by jiychoi           #+#    #+#             */
-/*   Updated: 2021/06/22 16:16:24 by jiychoi          ###   ########.fr       */
+/*   Updated: 2021/06/23 17:09:40 by jiychoi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,13 +31,13 @@ static void	server_get_string_if(int signo, int len_bit)
 	if (signo == SIGUSR1)
 	{
 		g_data_receive.char_temp |= 0;
-		if (len_bit < 8)
+		if (len_bit < 7)
 			g_data_receive.char_temp <<= 1;
 	}
 	else if (signo == SIGUSR2)
 	{
 		g_data_receive.char_temp |= 1;
-		if (len_bit < 8)
+		if (len_bit < 7)
 			g_data_receive.char_temp <<= 1;
 	}
 }
@@ -48,15 +48,15 @@ void		server_get_string(int signo, siginfo_t *info, void *ptr)
 	static int				len_byte;
 
 	ptr = (char *)ptr;
+	if (len_bit == 8)
+	{
+		len_bit = 0;
+		g_data_receive.str[len_byte++] = g_data_receive.char_temp;
+		g_data_receive.char_temp = 0;
+	}
 	if (len_byte < g_data_receive.length && g_data_receive.pid == info->si_pid)
 	{
-		if (len_bit == 8)
-		{
-			len_bit = 0;
-			g_data_receive.str[len_byte++] = g_data_receive.char_temp;
-			g_data_receive.char_temp = 0;
-		}
-		server_get_string_if(signo, len_bit);
+		server_get_string_if(signo, len_bit++);
 		kill(g_data_receive.pid, SIGUSR1);
 	}
 	if (len_byte == g_data_receive.length && g_data_receive.pid == info->si_pid)
@@ -66,8 +66,6 @@ void		server_get_string(int signo, siginfo_t *info, void *ptr)
 		sigaction(SIGUSR1, &sigact_srv_try_connect, 0);
 		sigaction(SIGUSR2, &sigact_srv_try_connect, 0);
 	}
-	if (g_data_receive.pid != info->si_pid)
-		kill(info->si_pid, SIGUSR2);
 }
 
 static void	server_get_length_if(int signo, int len_bit)
@@ -75,14 +73,14 @@ static void	server_get_length_if(int signo, int len_bit)
 	if (signo == SIGUSR1)
 	{
 		g_data_receive.length |= 0;
-		if (len_bit < 32)
+		if (len_bit < 31)
 			g_data_receive.length <<= 1;
 		kill(g_data_receive.pid, SIGUSR1);
 	}
 	else if (signo == SIGUSR2)
 	{
 		g_data_receive.length |= 1;
-		if (len_bit < 32)
+		if (len_bit < 31)
 			g_data_receive.length <<= 1;
 		kill(g_data_receive.pid, SIGUSR1);
 	}
@@ -94,57 +92,47 @@ void		server_get_length(int signo, siginfo_t *info, void *ptr)
 	char		*str_ptr;
 
 	ptr = (char *)ptr;
-	if (len_bit++ < 32 && g_data_receive.pid == info->si_pid)
-		server_get_length_if(signo, len_bit);
+	if (len_bit < 32 && g_data_receive.pid == info->si_pid)
+		server_get_length_if(signo, len_bit++);
 	if (len_bit == 32 && g_data_receive.pid == info->si_pid)
 	{
 		len_bit = 0;
-		sigaction(SIGUSR1, &sigact_srv_string, 0);
-		sigaction(SIGUSR2, &sigact_srv_string, 0);
 		str_ptr = (char *)malloc(sizeof(char) * (g_data_receive.length + 1));
 		if (!str_ptr)
+		{
+			ft_putstr_fd("Malloc failed. Exit server.\n", 1);
 			exit(1);
+		}
 		g_data_receive.str = str_ptr;
+		sigaction(SIGUSR1, &sigact_srv_string, 0);
+		sigaction(SIGUSR2, &sigact_srv_string, 0);
 		kill(g_data_receive.pid, SIGUSR1);
 	}
-	if (g_data_receive.pid != info->si_pid)
-		kill(info->si_pid, SIGUSR2);
 }
 
 void		server_connect(int signo, siginfo_t *info, void *ptr)
 {
 	ptr = (char *)ptr;
-	if (signo == SIGUSR1)
+	if (signo == SIGUSR1 && g_data_receive.pid == info->si_pid)
 	{
-		if (g_data_receive.pid == info->si_pid)
-		{
-			sigaction(SIGUSR1, &sigact_srv_length, 0);
-			sigaction(SIGUSR2, &sigact_srv_length, 0);
-			kill(g_data_receive.pid, SIGUSR1);
-		}
-		else
-			kill(info->si_pid, SIGUSR2);
+		sigaction(SIGUSR1, &sigact_srv_length, 0);
+		sigaction(SIGUSR2, &sigact_srv_length, 0);
+		kill(g_data_receive.pid, SIGUSR1);
 	}
 }
 
 void		server_try_connect(int signo, siginfo_t *info, void *ptr)
 {
 	ptr = (char *)ptr;
-	if (signo == SIGUSR1)
+	if (signo == SIGUSR1 && g_data_receive.pid == 0)
 	{
-		if (g_data_receive.pid == 0)
-		{
-			g_data_receive.pid = info->si_pid;
-			sigaction(SIGUSR1, &sigact_srv_try_connect, 0);
-			ft_putstr_fd("Connected with pid ", 1);
-			ft_putnbr_fd(g_data_receive.pid, 1);
-			ft_putchar_fd('\n', 1);
-			sigaction(SIGUSR1, &sigact_srv_connect, 0);
-			sigaction(SIGUSR2, &sigact_srv_connect, 0);
-			kill(g_data_receive.pid, SIGUSR1);
-		}
-		else
-			kill(info->si_pid, SIGUSR2);
+		g_data_receive.pid = info->si_pid;
+		ft_putstr_fd("Connected with pid ", 1);
+		ft_putnbr_fd(g_data_receive.pid, 1);
+		ft_putchar_fd('\n', 1);
+		sigaction(SIGUSR1, &sigact_srv_connect, 0);
+		sigaction(SIGUSR2, &sigact_srv_connect, 0);
+		kill(g_data_receive.pid, SIGUSR1);
 	}
 }
